[feat] Coupon Entity, Repository
- Coupon Entity 구현
- Coupon Repository 구현

+++ src/main/java/shop/sajotuna/order/coupon/domain/CouponSpecificCategory.java
@@ -14,10 +14,10 @@ import lombok.Setter;
 public class CouponSpecificCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int categoryCouponId;
+    private Long categoryCouponId;
 
     @Column(nullable = false)
-    private int categoryId;
+    private Integer categoryId;
 
     @ManyToOne
     @JoinColumn(name = "coupon_id", nullable = false)
